+++ packages/sync-core/tsup.config.ts
@@ -3,7 +3,8 @@ import { defineConfig } from 'tsup';
 export default defineConfig({
   entry: ['src/index.ts'],
   format: ['esm', 'cjs'],
-  dts: true,
+  tsconfig: 'tsconfig.build.json',
+  dts: { tsconfig: 'tsconfig.build.json' },
   sourcemap: true,
   clean: true,
 });
