+++ packages/sync-providers/package.json
@@ -2,21 +2,8 @@
   "name": "@sp/sync-providers",
   "version": "1.0.0",
   "description": "Framework-agnostic sync provider contracts for Super Productivity sync",
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
   "sideEffects": false,
   "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/index.d.mts",
-        "default": "./dist/index.mjs"
-      },
-      "require": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
-      }
-    },
     "./dropbox": {
       "import": {
         "types": "./dist/dropbox.d.mts",
@@ -129,8 +116,10 @@
     }
   },
   "dependencies": {
-    "@sp/sync-core": "*",
     "hash-wasm": "^4.12.0"
+  },
+  "peerDependencies": {
+    "@sp/sync-core": "*"
   },
   "scripts": {
     "build": "tsup",
@@ -139,6 +128,7 @@
     "test:watch": "vitest"
   },
   "devDependencies": {
+    "@sp/sync-core": "*",
     "@xmldom/xmldom": "^0.8.12",
     "tsup": "^8.0.0",
     "typescript": "^5.0.0",
