+++ packages/sync-providers/tests/file-based/local-file/local-file-sync-electron.spec.ts
@@ -2,12 +2,12 @@ import { describe, expect, it, vi } from 'vitest';
 import { NOOP_SYNC_LOGGER } from '@sp/sync-core';
 import {
   LocalFileSyncElectron,
-  type FileAdapter,
   type LocalFileSyncElectronDeps,
   type LocalFileSyncPrivateCfg,
   PROVIDER_ID_LOCAL_FILE,
-  type SyncCredentialStorePort,
-} from '../../../src';
+} from '../../../src/local-file';
+import type { FileAdapter } from '../../../src/file-based';
+import type { SyncCredentialStorePort } from '../../../src/credential-store';
 
 const fakeStore = (
   initial: LocalFileSyncPrivateCfg | null,
