+++ packages/sync-providers/tsup.config.ts
@@ -2,7 +2,6 @@ import { defineConfig } from 'tsup';
 
 export default defineConfig({
   entry: [
-    'src/index.ts',
     'src/dropbox.ts',
     'src/webdav.ts',
     'src/local-file.ts',
@@ -16,7 +15,8 @@ export default defineConfig({
     'src/provider-types.ts',
   ],
   format: ['esm', 'cjs'],
-  dts: true,
+  tsconfig: 'tsconfig.build.json',
+  dts: { tsconfig: 'tsconfig.build.json' },
   sourcemap: true,
   clean: true,
 });
